refactor(conversations): reuse ticket filter definitions
Generated-By: PostHog Code
Task-Id: 6570b974-9078-4221-8c59-968b15767487
