+++ products/dashboards/frontend/widgets/conversations/conversationsWidgetConfigValidation.ts
@@ -11,35 +11,9 @@ import { fieldErrorsFromZodError, parseWidgetConfig } from '../widgetConfigValid
 
 export type ConversationsTicketStatus = NonNullable<ConversationsRecentTicketsWidgetConfig['status']>
 export type ConversationsTicketPriority = NonNullable<ConversationsRecentTicketsWidgetConfig['priorities']>[number]
-export type ConversationsTicketChannel = NonNullable<ConversationsRecentTicketsWidgetConfig['channel']>
 export type ConversationsTicketAssignee = NonNullable<ConversationsRecentTicketsWidgetConfig['assignees']>[number]
 type ConversationsWidgetFormField = keyof z.infer<typeof conversationsRecentTicketsWidgetFormSchema>
 export type ConversationsWidgetFieldErrors = Partial<Record<ConversationsWidgetFormField, string>>
-
-export const CONVERSATIONS_TICKET_STATUS_OPTIONS: { value: ConversationsTicketStatus; label: string }[] = [
-    { value: 'all', label: 'All statuses' },
-    { value: 'new', label: 'New' },
-    { value: 'open', label: 'Open' },
-    { value: 'pending', label: 'Pending' },
-    { value: 'on_hold', label: 'On hold' },
-    { value: 'resolved', label: 'Resolved' },
-]
-
-export const CONVERSATIONS_TICKET_PRIORITY_OPTIONS: { key: ConversationsTicketPriority; label: string }[] = [
-    { key: 'low', label: 'Low' },
-    { key: 'medium', label: 'Medium' },
-    { key: 'high', label: 'High' },
-    { key: 'critical', label: 'Critical' },
-]
-
-export const CONVERSATIONS_TICKET_CHANNEL_OPTIONS: { value: ConversationsTicketChannel; label: string }[] = [
-    { value: 'all', label: 'All channels' },
-    { value: 'widget', label: 'Widget' },
-    { value: 'slack', label: 'Slack' },
-    { value: 'teams', label: 'Microsoft Teams' },
-    { value: 'email', label: 'Email' },
-    { value: 'github', label: 'GitHub' },
-]
 
 export function parseConversationsWidgetConfig(
     config: Record<string, unknown>
